+++ test/cloudex/cloudex_test.exs
@@ -10,7 +10,8 @@ defmodule CloudexTest do
   doctest Cloudex
 
   test "file not found" do
-    assert {:error, "File /non/existing/file.jpg does not exist."} = Cloudex.upload("/non/existing/file.jpg")
+    assert {:error, "File /non/existing/file.jpg does not exist."} =
+             Cloudex.upload("/non/existing/file.jpg")
   end
 
   test "upload single image file" do
@@ -31,27 +32,43 @@ defmodule CloudexTest do
 
   test "upload image url" do
     use_cassette "test_upload_url with both http and https" do
-      assert {:ok, %Cloudex.UploadedImage{}} = Cloudex.upload("http://cdn.mhpbooks.com/uploads/2014/10/shutterstock_172896005.jpg")
-      assert {:ok, %Cloudex.UploadedImage{}} = Cloudex.upload("https://cdn.mhpbooks.com/uploads/2014/10/shutterstock_172896005.jpg")
+      assert {:ok, %Cloudex.UploadedImage{}} =
+               Cloudex.upload(
+                 "http://cdn.mhpbooks.com/uploads/2014/10/shutterstock_172896005.jpg"
+               )
+
+      assert {:ok, %Cloudex.UploadedImage{}} =
+               Cloudex.upload(
+                 "https://cdn.mhpbooks.com/uploads/2014/10/shutterstock_172896005.jpg"
+               )
     end
   end
 
   test "mixed files / urls" do
     use_cassette "test_upload_mixed" do
-      assert  [
-                {:ok, %Cloudex.UploadedImage{}},
-                {:ok, %Cloudex.UploadedImage{}},
-                {:error, "File nonexistent.png does not exist."}
-              ] = Cloudex.upload(["./test/assets/test.jpg", "nonexistent.png", "https://cdn.mhpbooks.com/uploads/2014/10/shutterstock_172896005.jpg"])
+      assert [
+               {:ok, %Cloudex.UploadedImage{}},
+               {:ok, %Cloudex.UploadedImage{}},
+               {:error, "File nonexistent.png does not exist."}
+             ] =
+               Cloudex.upload([
+                 "./test/assets/test.jpg",
+                 "nonexistent.png",
+                 "https://cdn.mhpbooks.com/uploads/2014/10/shutterstock_172896005.jpg"
+               ])
     end
   end
 
   test "upload with tags" do
     use_cassette "test_upload_with_tags" do
       tags = ["foo", "bar"]
-      {:ok, %Cloudex.UploadedImage{tags: ^tags}} = Cloudex.upload(["./test/assets/test.jpg"], %{tags: Enum.join(tags, ",")})
+
+      {:ok, %Cloudex.UploadedImage{tags: ^tags}} =
+        Cloudex.upload(["./test/assets/test.jpg"], %{tags: Enum.join(tags, ",")})
+
       # or simply
-      {:ok, %Cloudex.UploadedImage{tags: ^tags}} = Cloudex.upload(["./test/assets/test.jpg"], %{tags: tags})
+      {:ok, %Cloudex.UploadedImage{tags: ^tags}} =
+        Cloudex.upload(["./test/assets/test.jpg"], %{tags: tags})
     end
   end
 
@@ -64,15 +81,18 @@ defmodule CloudexTest do
 
   test "delete image with public id" do
     use_cassette "test_delete" do
-      assert {:ok, %Cloudex.DeletedImage{public_id: "rurwrndtvgzfajljllnr"}} = Cloudex.delete("rurwrndtvgzfajljllnr")
+      assert {:ok, %Cloudex.DeletedImage{public_id: "rurwrndtvgzfajljllnr"}} =
+               Cloudex.delete("rurwrndtvgzfajljllnr")
     end
   end
 
   test "delete image with invalid public id" do
     use_cassette "test_delete_invalid" do
-      assert {:ok, %Cloudex.DeletedImage{public_id: "thisIsABogusId"}} = Cloudex.delete("thisIsABogusId")
+      assert {:ok, %Cloudex.DeletedImage{public_id: "thisIsABogusId"}} =
+               Cloudex.delete("thisIsABogusId")
     end
   end
+
   test "delete images from a list of public id's" do
     use_cassette "test_delete_list" do
       assert [
@@ -82,10 +102,10 @@ defmodule CloudexTest do
     end
   end
 
-
   test "create a uploaded image from a map" do
     {:ok, data} = Poison.decode(File.read!("./test/cloudinary_response.json"))
     result = Cloudex.CloudinaryApi.json_result_to_struct(data, "http://example.org/test.jpg")
+
     assert %Cloudex.UploadedImage{
              bytes: 22659,
              created_at: "2015-11-27T10:02:23Z",
@@ -95,13 +115,15 @@ defmodule CloudexTest do
              original_filename: "test",
              public_id: "i2nruesgu4om3w9mtk1z",
              resource_type: "image",
-             secure_url: "https://d1vibqt9pdnk2f.cloudfront.net/image/upload/v1448618543/i2nruesgu4om3w9mtk1z.jpg",
+             secure_url:
+               "https://d1vibqt9pdnk2f.cloudfront.net/image/upload/v1448618543/i2nruesgu4om3w9mtk1z.jpg",
              signature: "77b447746476c82bb4921fdea62a9227c584974b",
              source: "http://example.org/test.jpg",
              tags: [],
              type: "upload",
-             url: "http://images.cloudassets.mobi/image/upload/v1448618543/i2nruesgu4om3w9mtk1z.jpg",
-             version: 1448618543,
+             url:
+               "http://images.cloudassets.mobi/image/upload/v1448618543/i2nruesgu4om3w9mtk1z.jpg",
+             version: 1_448_618_543,
              width: 250
            } = result
   end